chore: move some stuff to src
