+++ scripts/fetch-ccns.js
@@ -2,30 +2,13 @@ import fs from "fs";
 import path from "path";
 import pMap from "p-map";
 
-import { getCCNContainer } from "./fetch-steps/get-ccn-container";
-import { embedCCNTexts } from "./fetch-steps/embed-ccn-texts";
-
-import { filterData } from "../src/filterData";
-import astify from "../src/astify";
-
 // fetch all conventions from index.json
 import conventions from "../data/index.json";
-
-const fetchCCN = id =>
-  getCCNContainer(id)
-    //.then(re => console.log("getContainer", re) || re)
-    .then(filterData) // first pass to reduce the number of text calls
-    .then(embedCCNTexts) // add textes attaches & salaires
-    .then(filterData) // second and final pass to trim and order additional texts
-    .then(astify) // convert to AST format
-    .then(checkCcn); // check validity
+import fetchCcn from "../src/fetchCcn";
 
 const strfy = data => JSON.stringify(data, null, 2);
 
 const checkCcn = ccn => {
-  if (ccn.data.num.length != 4) {
-    throw new Error(`ccn ${ccn.data.id}: invalid num "${ccn.data.num}"`);
-  }
   if (ccn.children.length === 0) {
     throw new Error(`ccn ${ccn.data.id}: no children"`);
   }
@@ -39,16 +22,18 @@ const fetchAllConventions = () =>
     convention => {
       const filePath = path.join(__dirname, `../data/${convention.id}.json`);
       console.log(`fetch ${convention.id}`);
-      return fetchCCN(convention.id).then(data => {
-        const exists = fs.existsSync(filePath);
-        const changed = strfy(exists && require(filePath)) !== strfy(data); // poor-man compare
-        if (changed) {
-          fs.writeFileSync(filePath, strfy(data));
-          console.log(`wrote ${filePath}`);
-        } else {
-          console.log(`skip ${filePath} (not changed)`);
-        }
-      });
+      return fetchCcn(convention.id)
+        .then(checkCcn)
+        .then(data => {
+          const exists = fs.existsSync(filePath);
+          const changed = strfy(exists && require(filePath)) !== strfy(data); // poor-man compare
+          if (changed) {
+            fs.writeFileSync(filePath, strfy(data));
+            console.log(`wrote ${filePath}`);
+          } else {
+            console.log(`skip ${filePath} (not changed)`);
+          }
+        });
       //  .catch(console.error);
     },
     { concurrency: 1, stopOnError: true }
